fix the bug, and promote the ui

+++ src/filesystem/service/FATService.java
@@ -22,7 +22,7 @@ public class FATService {
 	}
 	//apply the space of block,set the FAT
 	public static void applyForBlock(int pre,int index,FATModel fat) {
-		fat.getTable()[pre] = (byte)index;
+		fat.getTable()[pre] = index;
 		fat.setFreeCount(fat.getFreeCount()-1);
 	}
 	//set the value of block
